First wave of major refactoring/reorganizing
